Exp01 benchmark

+++ backboard/benchmark/benchmark.py
@@ -1,3 +1,5 @@
+import os
+import warnings
 from collections import defaultdict
 
 import pandas
@@ -6,7 +8,7 @@ from torch.optim import SGD
 from backboard.benchmark.utils import get_train_size
 from backboard.cockpit_plotter import CockpitPlotter
 from backboard.quantities import Time
-from backboard.runners.scheduled_runner import ScheduleCockpitRunner
+from backboard.runners.scheduled_runner import _ScheduleCockpitRunner
 from deepobs.pytorch.config import set_default_device
 
 
@@ -36,7 +38,33 @@ def _check_quantities(quantities):
             )
 
 
-def _make_runner(quantities):
+def _runner_stop_after(step=None):
+    class BenchmarkRunner(_ScheduleCockpitRunner):
+        """Runner with disabled computation DeepOBS' additional metrics."""
+
+        STOP_AFTER = step
+
+        def _maybe_stop_iteration(self, global_step, batch_count):
+            """Don't interrupt."""
+            if self.STOP_AFTER is None:
+                return
+
+            if getattr(self, "_already_stopped", False):
+                raise RuntimeError("Runner must stop at last epoch.")
+
+            if global_step > self.STOP_AFTER:
+                warnings.warn(f"BenchmarkRunner stopping epoch at step {global_step}")
+                self._already_stopped = True
+                raise StopIteration
+
+        def _should_eval(self):
+            """Disable DeepOBS' evaluation of test/train/valid losses and accuracies."""
+            return False
+
+    return BenchmarkRunner
+
+
+def _make_runner(quantities, steps):
     """Return a DeepOBS runner that tracks the specified quantities."""
     optimizer_class = SGD
     hyperparams = {
@@ -45,8 +73,18 @@ def _make_runner(quantities):
         "nesterov": {"type": bool, "default": False},
     }
 
-    return ScheduleCockpitRunner(
-        optimizer_class, hyperparams, quantities=quantities, plot=False
+    def plot_schedule(global_step):
+        """Never plot."""
+        return False
+
+    runner_cls = _runner_stop_after(step=steps)
+
+    return runner_cls(
+        optimizer_class,
+        hyperparams,
+        quantities=quantities,
+        plot=False,
+        plot_schedule=plot_schedule,
     )
 
 
@@ -91,7 +129,7 @@ def run_benchmark(testproblem, quantities, steps, random_seed):
     _check_timer(quantities, steps)
     _check_quantities(quantities)
 
-    runner = _make_runner(quantities)
+    runner = _make_runner(quantities, steps)
     num_epochs = _get_num_epochs(runner, testproblem, steps)
 
     runner.run(
@@ -143,7 +181,7 @@ def _compute_steps(steps, track_events, track_interval):
         return steps
 
 
-def benchmark(
+def benchmark(  # noqa: C901
     testproblems,
     configs,
     track_intervals,
@@ -168,7 +206,17 @@ def benchmark(
         "device",
         "time_per_step",
     ]
-    data = pandas.DataFrame(columns=columns)
+
+    if savefile is not None and os.path.exists(savefile):
+        # try loading
+        data = pandas.read_csv(savefile, comment="#", index_col=[0])
+        assert set(data.columns) == set(columns), (
+            f"Loaded file {savefile} has inconsistent columns:"
+            + f"\n\tFound:   {list(data.columns)}"
+            + f"\n\tRequire: {columns}"
+        )
+    else:
+        data = pandas.DataFrame(columns=columns)
 
     for device in devices:
         set_default_device(device)
@@ -180,11 +228,6 @@ def benchmark(
 
                         this_steps = _compute_steps(steps, track_events, track_interval)
 
-                        quantities = [q(track_interval=track_interval) for q in config]
-                        runtime = run_benchmark(
-                            testproblem, quantities, this_steps, random_seed
-                        )
-
                         run_data = {
                             "testproblem": testproblem,
                             "quantities": name,
@@ -192,17 +235,49 @@ def benchmark(
                             "steps": this_steps,
                             "random_seed": random_seed,
                             "device": device,
-                            "time_per_step": runtime,
                         }
-                        data = data.append(run_data, ignore_index=True)
 
-                        if savefile is not None:
-                            with open(savefile, "w") as f:
-                                if header is not None:
-                                    header_comment = "\n".join(
-                                        "# " + line for line in header.splitlines()
-                                    )
-                                    f.write(header_comment + "\n")
-                                data.to_csv(f)
+                        run_exists = (
+                            (data["testproblem"] == testproblem)
+                            & (data["quantities"] == name)
+                            & (data["steps"] == this_steps)
+                            & (data["track_interval"] == track_interval)
+                            & (data["random_seed"] == random_seed)
+                            & (data["device"] == device)
+                        ).any()
+
+                        if run_exists:
+                            print(f"Setting already exists and is skipped: {run_data}")
+                        else:
+                            print(f"Running setting: {run_data}")
+
+                            def track_schedule(global_step):
+                                return (
+                                    global_step >= 0
+                                    and global_step % track_interval == 0
+                                )
+
+                            quantities = [
+                                q(
+                                    track_schedule=track_schedule,
+                                    # verbose=True,
+                                )
+                                for q in config
+                            ]
+                            runtime = run_benchmark(
+                                testproblem, quantities, this_steps, random_seed
+                            )
+                            run_data["time_per_step"] = runtime
+
+                            data = data.append(run_data, ignore_index=True)
+
+                            if savefile is not None:
+                                with open(savefile, "w") as f:
+                                    if header is not None:
+                                        header_comment = "\n".join(
+                                            "# " + line for line in header.splitlines()
+                                        )
+                                        f.write(header_comment + "\n")
+                                    data.to_csv(f)
 
     return data
